+++ Setup.php
@@ -8,6 +8,9 @@ use XF\AddOn\StepRunnerUninstallTrait;
 use XF\AddOn\StepRunnerUpgradeTrait;
 use XF\Db\Schema\Alter;
 
+/**
+ * @version 1.1.1
+ */
 class Setup extends AbstractSetup
 {
     use StepRunnerInstallTrait;
@@ -19,7 +22,7 @@ class Setup extends AbstractSetup
         'xf_rm_resource_version'
     ];
 
-    public function installStep1()
+    public function installStep1() : void
     {
         $sm = $this->schemaManager();
 
@@ -27,13 +30,13 @@ class Setup extends AbstractSetup
         {
             $sm->alterTable($tableName, function(Alter $table)
             {
-                $table->addColumn('user_id', 'int');
-                $table->addColumn('username', 'varchar', 100)->setDefault('');
+                $table->addColumn('tck_uar_user_id', 'int');
+                $table->addColumn('tck_uar_username', 'varchar', 100)->setDefault('');
             });
         }
     }
 
-    public function uninstallStep1()
+    public function upgrade1010170Step1() : void
     {
         $sm = $this->schemaManager();
 
@@ -41,7 +44,23 @@ class Setup extends AbstractSetup
         {
             $sm->alterTable($tableName, function(Alter $table)
             {
-                $table->dropColumns(['user_id', 'username']);
+                foreach (['user_id', 'username'] AS $column)
+                {
+                    $table->renameColumn($column, "tck_uar_{$column}");
+                }
+            });
+        }
+    }
+
+    public function uninstallStep1() : void
+    {
+        $sm = $this->schemaManager();
+
+        foreach ($this->tables AS $tableName)
+        {
+            $sm->alterTable($tableName, function(Alter $table)
+            {
+                $table->dropColumns(['tck_uar_user_id', 'tck_uar_username']);
             });
         }
     }