+++ XFRM/Entity/ResourceUpdate.php
@@ -2,7 +2,6 @@
 
 namespace TickTackk\UpdateAnyResource\XFRM\Entity;
 
-use TickTackk\UpdateAnyResource\Globals;
 use TickTackk\UpdateAnyResource\XFRM\Entity\XF21\ResourceItem as XF21ResourceItemEntity;
 use TickTackk\UpdateAnyResource\XFRM\Entity\XF22\ResourceItem as XF22ResourceItemEntity;
 use XF\Mvc\Entity\Structure;
@@ -11,12 +10,11 @@ use XF\Entity\User as UserEntity;
 
 /**
  * COLUMNS
- * @property int user_id
- * @property int user_id_
- * @property string username
+ * @property int $tck_uar_user_id
+ * @property string $tck_uar_username
  *
  * RELATIONS
- * @property UserEntity User
+ * @property UserEntity $User
  * @property ResourceItemTrait|XF21ResourceItemEntity|XF22ResourceItemEntity $Resource
  */
 class ResourceUpdate extends XFCP_ResourceUpdate
@@ -34,7 +32,7 @@ class ResourceUpdate extends XFCP_ResourceUpdate
             return false;
         }
 
-        return $resource->runActionForTckUpdateAnyResource($this->user_id, function () use($error)
+        return $resource->runActionForTckUpdateAnyResource($this->tck_uar_user_id, function () use($error)
         {
             return parent::canEdit($error);
         });
@@ -54,7 +52,7 @@ class ResourceUpdate extends XFCP_ResourceUpdate
             return false;
         }
 
-        return $resource->runActionForTckUpdateAnyResource($this->user_id, function () use($type, $error)
+        return $resource->runActionForTckUpdateAnyResource($this->tck_uar_user_id, function () use($type, $error)
         {
             return parent::canDelete($type, $error);
         });
@@ -67,7 +65,7 @@ class ResourceUpdate extends XFCP_ResourceUpdate
     {
         return (
             parent::canSendModeratorActionAlert()
-            && (\XF::visitor()->user_id !== $this->user_id)
+            && (\XF::visitor()->user_id !== $this->tck_uar_user_id)
         );
     }
 
@@ -80,14 +78,16 @@ class ResourceUpdate extends XFCP_ResourceUpdate
     {
         $structure = parent::getStructure($structure);
 
-        $structure->columns['user_id'] = ['type' => self::UINT, 'required' => true];
-        $structure->columns['username'] = ['type' => self::STR, 'maxLength' => 50,
+        $structure->columns['tck_uar_user_id'] = ['type' => self::UINT, 'required' => true];
+        $structure->columns['tck_uar_username'] = ['type' => self::STR, 'maxLength' => 50,
             'required' => 'please_enter_valid_name'
         ];
         $structure->relations['User'] = [
             'entity' => 'XF:User',
             'type' => self::TO_ONE,
-            'conditions' => 'user_id',
+            'conditions' => [
+                ['user_id', '=', '$tck_uar_user_id']
+            ],
             'primary' => true
         ];