+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 - **Fix:** Since 2.2, moderator action alert cannot be sent if the user is a team member (#6)
 - **Fix:** Since 2.2, team members are unable to post new resource updates (#5)
 - **Fix:** User who is not team member and made the release cannot download without permission (#7)
+- **Fix:** Username/user id columns added in resource update/version are not prefixed (#8)
 
 ## 1.1.0 (`1000070`)